requirements.txt
# Core ML and Data Processing
numpy>=1.21.0
pandas>=1.3.0
torch>=2.0.0
torchvision>=0.15.0
scikit-learn>=1.0.0

# Geospatial Data Processing
rasterio>=1.3.0
gdal>=3.4.0
geopandas>=0.12.0
shapely>=2.0.0
pyproj>=3.4.0
fiona>=1.9.0

# Computer Vision
opencv-python>=4.5.0
pillow>=9.0.0
albumentations>=1.3.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.10.0

# Utilities
tqdm>=4.62.0
pyyaml>=6.0
requests>=2.26.0
python-dotenv>=0.19.0

# Satellite Data APIs
sentinelsat>=1.1.0
earthengine-api>=0.1.340

# Distributed Computing (optional)
dask[complete]>=2023.1.0

# API Framework (optional)
fastapi>=0.95.0
uvicorn[standard]>=0.20.0
pydantic>=2.0.0
python-multipart>=0.0.5

# MLOps (optional)
mlflow>=2.1.0
optuna>=3.1.0

# Testing and Development
pytest>=7.0.0
pytest-cov>=3.0.0
black>=22.0.0
flake8>=4.0.0
mypy>=0.950
jupyter>=1.0.0
ipython>=8.0.0
